Solved sorting of negative numbers

+++ libft/src/linked_lists/ft_lstsorted.c
@@ -1,10 +1,12 @@
 #include "libft.h"
+#include "stdio.h"
 
 int ft_lstsorted(t_list *lst)
 {
 	 while (lst->next)
 	 {
-		if (lst->data > lst->next->data)
+		//printf("%li > %li\n", (long int)lst->data, (long int)lst->next->data);
+		if ((long int)lst->data > (long int)lst->next->data)
 			return (0);
 		lst = lst->next;
 	 }
